Does it check for a minimum version of libraries? It looks like it ch… (#162)
* Does it check for a minimum version of libraries? It looks like it checks for an exact match.
Also, go ahead and put pyParsing, scons, and gitPython in. Are there others that are not installed by default?
Have you tested? Remove one and see if it reinstalls.
(DONE)

* Does it check for a minimum version of libraries? It looks like it checks for an exact match.
Also, go ahead and put pyParsing, scons, and gitPython in. Are there others that are not installed by default?
Have you tested? Remove one and see if it reinstalls.
(DONE)

Co-authored-by: Kiran <kiran.malvi@mindinventory.com>
Co-authored-by: Bruce Long <qstream@gmail.com>